Fixed unit tests

+++ scripts/activities/email-file.sh
@@ -37,12 +37,15 @@ set -xe
 BASENAME=$(basename ${FILENAME} .gz)
 ATTACHMENTS=""
 n=0
+CURRENT_DIR=$(pwd)
 
 for dir in ${INPUT1_STAGING_DIR} ${INPUT2_STAGING_DIR} ${INPUT3_STAGING_DIR} ${INPUT4_STAGING_DIR} ${INPUT5_STAGING_DIR} ${INPUT6_STAGING_DIR} ${INPUT7_STAGING_DIR} ${INPUT8_STAGING_DIR} ${INPUT9_STAGING_DIR} ${INPUT10_STAGING_DIR}; do
   # Create a temporary working dir for this file
-  WORKING_DIR=$(mktemp -d)
+  if [ -z "${WORKING_DIR}" ]; then
+    WORKING_DIR=$(mktemp)
+    add_on_exit rm -rf ${WORKING_DIR}
+  fi
   cd ${WORKING_DIR}
-  add_on_exit rm -rf ${WORKING_DIR}
 
   # Decompress the files if required
   find ${dir} -name \*.gz | xargs gunzip
@@ -65,6 +68,8 @@ for dir in ${INPUT1_STAGING_DIR} ${INPUT2_STAGING_DIR} ${INPUT3_STAGING_DIR} ${I
   fi
 
   n=$((n+1))
+
+  cd ${CURRENT_DIR}
 done
 
 ${INSTALL_MAILER}
